connexio elastic ok + funcions elastic

+++ src/main/java/com/RubenJimenez/TFG/service/ProductService.java
@@ -1,17 +1,19 @@
 package com.RubenJimenez.TFG.service;
 
 import com.RubenJimenez.TFG.models.Product;
-import com.RubenJimenez.TFG.config.*;
+
 import com.RubenJimenez.TFG.repo.ProductRepo;
+
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class ProductService {
 
-    @Autowired
-    private RestHighLevelClient client;
+
     @Autowired
     private ProductRepo productRepo;
 
@@ -25,16 +27,22 @@ public class ProductService {
         return productRepo.save(prod);
     }
 
-    public Product updateProduct(Product prod,int id){
+    /*public Product updateProduct(Product prod,int id){
 
         Product prod1 = productRepo.findById(id).get();
         prod1.setPrice(prod.getPrice());
         return prod1;
-    }
+    }*/
 
-    public int deleteProduct(int id){
-        productRepo.deleteById(id);
-        return 1;
+    public Optional<Product> deleteProduct(int id) {
+
+        Optional<Product> product = productRepo.findById(id);
+
+        if (product.isPresent()) {
+            productRepo.deleteById(id);
+        }
+
+        return product;
     }
 
 }
